spacex_pipeline/requirements.txt
apache-airflow-providers-postgres
psycopg2-binary
pandas
requests
SQLAlchemy